extGIA: Fixed ImagesDAO and updated tests

+++ extGIA/src/es/icarto/gvsig/extgia/forms/images/ImagesDAO.java
@@ -16,6 +16,9 @@ public class ImagesDAO {
 
     public void insertImageIntoDb(Connection connection, String schema, String tablename,
 	    String pkField, String pkValue, BufferedImage image, boolean update) throws SQLException, IOException {
+	if (schema == null || tablename == null || pkField == null || pkValue == null || image == null) {
+	    return;
+	}
 	byte[] imageBytes = ImageUtils.convertImageToBytea(image);
 	PreparedStatement statement;
 	if (update) {
@@ -48,6 +51,9 @@ public class ImagesDAO {
     public byte[] readImageFromDb(Connection connection, String schema, String tablename,
 	    String pkField, String pkValue) throws SQLException {
 	PreparedStatement statement = null;
+	if (schema == null || tablename == null || pkField == null || pkValue == null) {
+	    return null;
+	}
 	try {
 	    statement = connection.prepareStatement("SELECT "
 		    + IMAGE_FIELDNAME + " FROM "
@@ -74,6 +80,9 @@ public class ImagesDAO {
     public void deleteImageFromDb(Connection connection, String schema, String tablename,
 	    String pkField, String pkValue) throws SQLException {
 	PreparedStatement statement = null;
+	if (schema == null || tablename == null || pkField == null || pkValue == null) {
+	    return;
+	}
 	try {
 	    statement = connection.prepareStatement("DELETE FROM "
 		    + schema + "." + tablename
